match default audio behavior to upstream/Mednafen
This seems like a similar situation to the BIOS ROM name issue from earlier this week: https://github.com/libretro/beetle-pcfx-libretro/pull/15

In porting to libretro, the various 'get setting' functions were simplified to return only the values needed for that particular core. The default value in standalone Mednafen for this setting is 1, but it wound up as 0 in the port. http://mednafen.fobby.net/documentation/pcfx.html#Settings+Reference

There are forums users who are experiencing the 'click' audio error which this PR should fix.